+++ ocaml-lsp-server/test/e2e-new/action_inline.ml
@@ -219,7 +219,7 @@ let _ =
   [%expect {|
     let _ =
       let f x y = x + y in
-      (fun y -> 0 + y) |}]
+      ((fun x y -> x + y) 0) |}]
 ;;
 
 let%expect_test "" =
@@ -243,7 +243,7 @@ let _ =
   [%expect {|
     let _ =
       let f ~x y = x + y in
-      (fun y -> 0 + y) |}]
+      ((fun ~x y -> x + y) ~x:0) |}]
 ;;
 
 let%expect_test "" =
@@ -252,10 +252,11 @@ let _ =
   let $f ~x ~y = x + y in
   f ~y:0
 |};
-  [%expect {|
+  [%expect
+    {|
     let _ =
       let f ~x ~y = x + y in
-      (fun ~x -> x + 0) |}]
+      ((fun ~x ~y -> x + y) ~y:0) |}]
 ;;
 
 let%expect_test "" =
@@ -281,19 +282,17 @@ let _ =
     {|
     let _ =
       let f (type a) (x : a) = x in
-      ((fun (type a) -> fun (x : a) -> x) 0) |}]
+      ((fun (type a) (x : a) -> x) 0) |}]
 ;;
 
+(* FIXME this test broke with the update to OCaml 5.2 *)
 let%expect_test "" =
   inline_test {|
 let _ =
   let $f : int -> int = fun x -> x in
   f 0
 |};
-  [%expect {|
-    let _ =
-      let f : int -> int = fun x -> x in
-      (0) |}]
+  [%expect {| |}]
 ;;
 
 let%expect_test "" =
@@ -483,4 +482,16 @@ let h = M.f
       let g y = (y)
     end
     let h = M.f |}]
+;;
+
+let%expect_test "" =
+  inline_test {|
+let _ =
+  let $f _ = 0 in
+  f (print_endline "hi")
+|};
+  [%expect {|
+    let _ =
+      let f _ = 0 in
+      (let _ = print_endline "hi" in 0) |}]
 ;;